+++ src/app/core/services/constants.service.ts
@@ -1,9 +1,5 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ConstantsService {
-
-  constructor() { }
-}
+export const CONSTANTS_SERVICE = {
+  App: 'TaskManager',
+  Ver: '1.0',
+  API_URL: 'http://...'
+};