+++ src/app/shared/shared.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
+import { ClickFontsDirective } from './directives/click-fonts.directive';
 
 
 
 @NgModule({
-  declarations: [HighlightDirective],
+  declarations: [HighlightDirective, ClickFontsDirective],
   exports: [
     HighlightDirective
   ],